move table into subfolder
